// resources/js/types/auth.ts
export type User = {
    id: number;
    name: string;
    email: string;
    avatar?: string;
    email_verified_at: string | null;
    two_factor_enabled?: boolean;
    created_at: string;
    updated_at: string;
    roleuser: Role;
    [key: string]: unknown;
};

export type Auth = {
    user: User;
};

export type TwoFactorSetupData = {
    svg: string;
    url: string;
};

export type TwoFactorSecretKey = {
    secretKey: string;
};

export type Role = {
    jenjang: string;
    nama_sekolah: string;
    npsn: string;
    slug: string;
    sekolah_id: number;
    sekolah_permission: string;
};